refactor: move utils to own file
